requirements.txt
alabaster==0.7.13
appdirs==1.4.4
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
black==23.1.0
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
docutils==0.19
esbonio==0.16.0
exceptiongroup==1.1.0
executing==1.2.0
fonttools==4.38.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
lsprotocol==2022.0.0a10
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pygls==1.0.0
Pygments==2.14.0
pynput==1.7.6
pyparsing==3.0.9
pyspellchecker==0.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.10.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
tweepy==4.12.1
twitter==1.19.6
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.12.1
